Expose createLoaderElement method from react-cosmos package #276
