+++ packages/react-cosmos/src/index.js
@@ -1,7 +1,9 @@
+import createLoaderElement from './create-loader-element';
 import startLoader from './start-loader';
 import startPlayground from './start-playground';
 
 module.exports = {
+  createLoaderElement,
   startLoader,
   startPlayground,
 };
